requirements.txt
python-telegram-bot==20.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
flask==2.3.2